# backend/ism-resource/src/main/java/io/github/turbopro/ism/resource/supplier/SupplierResourceController.java
package io.github.turbopro.ism.resource.supplier;
import io.github.turbopro.ism.common.api.ApiResponse;import io.github.turbopro.ism.common.infrastructure.authorization.RequiresPermission;import io.github.turbopro.ism.common.infrastructure.web.ApiResponseFactory;import jakarta.validation.Valid;import jakarta.validation.constraints.*;import org.springframework.validation.annotation.Validated;import org.springframework.web.bind.annotation.*;
@Validated @RestController @RequestMapping("/api/resources") public class SupplierResourceController{
 private final SupplierResourceService service;private final ApiResponseFactory responses;public SupplierResourceController(SupplierResourceService s,ApiResponseFactory r){service=s;responses=r;}
 @GetMapping("/persons")@RequiresPermission("resource:person:view")ApiResponse<SupplierResourceModels.PersonPage>persons(@RequestParam(defaultValue="")@Size(max=100)String keyword,@RequestParam(defaultValue="")String status,@RequestParam(defaultValue="0")@Min(0)int page,@RequestParam(defaultValue="20")@Min(1)@Max(100)int size){return responses.success(service.persons(keyword,status,page,size));}
 @PostMapping("/persons")@RequiresPermission("resource:person:manage")ApiResponse<SupplierResourceModels.PersonView>createPerson(@Valid@RequestBody SupplierResourceModels.SavePerson c){return responses.success(service.createPerson(c));}
 @PutMapping("/persons/{id}")@RequiresPermission("resource:person:manage")ApiResponse<SupplierResourceModels.PersonView>updatePerson(@PathVariable long id,@Valid@RequestBody SupplierResourceModels.SavePerson c){return responses.success(service.updatePerson(id,c));}
 @PostMapping("/persons/{id}/status")@RequiresPermission("resource:person:status")ApiResponse<SupplierResourceModels.PersonView>personStatus(@PathVariable long id,@Valid@RequestBody SupplierResourceModels.PersonStatusCommand c){return responses.success(service.personStatus(id,c));}
 @GetMapping("/assets")@RequiresPermission("resource:asset:view")ApiResponse<SupplierResourceModels.AssetPage>assets(@RequestParam(defaultValue="")@Size(max=100)String keyword,@RequestParam(defaultValue="")String status,@RequestParam(defaultValue="0")@Min(0)int page,@RequestParam(defaultValue="20")@Min(1)@Max(100)int size){return responses.success(service.assets(keyword,status,page,size));}
 @PostMapping("/assets")@RequiresPermission("resource:asset:manage")ApiResponse<SupplierResourceModels.AssetView>createAsset(@Valid@RequestBody SupplierResourceModels.SaveAsset c){return responses.success(service.createAsset(c));}
 @PutMapping("/assets/{id}")@RequiresPermission("resource:asset:manage")ApiResponse<SupplierResourceModels.AssetView>updateAsset(@PathVariable long id,@Valid@RequestBody SupplierResourceModels.SaveAsset c){return responses.success(service.updateAsset(id,c));}
 @PostMapping("/assets/{id}/status")@RequiresPermission("resource:asset:status")ApiResponse<SupplierResourceModels.AssetView>assetStatus(@PathVariable long id,@Valid@RequestBody SupplierResourceModels.AssetStatusCommand c){return responses.success(service.assetStatus(id,c));}
 @PostMapping("/assets/{id}/handover")@RequiresPermission("resource:asset:status")ApiResponse<SupplierResourceModels.AssetView>handoverAsset(@PathVariable long id,@Valid@RequestBody SupplierResourceModels.AssetHandoverCommand c){return responses.success(service.handoverAsset(id,c));}
}